refactor(tests): realocate folder
