+++ legacy_tests/Browser/AddError/Component.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Tests;
+namespace LegacyTests\Browser\AddError;
 
 use Illuminate\Support\Facades\View;
 use Livewire\Component as BaseComponent;
