+++ legacy_tests/Browser/AddError/Test.php
