+++ legacy_tests/Browser/AddError/view.blade.php
